refactored migrations, models and schemata
