+++ src/models/Deputy.js
@@ -1,19 +1,4 @@
-import mongoose, { Schema } from 'mongoose';
-import DeputyContact from './Schemas/DeputyContact';
-
-const DeputySchema = new Schema(
-  {
-    webId: { type: String },
-    imgURL: { type: String },
-    name: { type: String },
-    party: { type: String },
-    job: { type: String },
-    biography: { type: String },
-    constituency: { type: String },
-    directCandidate: { type: Boolean },
-    contact: { type: DeputyContact },
-  },
-  { timestamps: true },
-);
+import mongoose from 'mongoose';
+import DeputySchema from './../migrations/3-schemas/Deputy';
 
 export default mongoose.model('Deputy', DeputySchema);
