+++ src/models/Procedure.js
@@ -1,93 +1,9 @@
-import mongoose, { Schema } from 'mongoose';
-
-import Document from './Schemas/Document';
-import procedureStates from '../config/procedureStates';
-
-const ProcedureSchema = new Schema(
-  {
-    procedureId: { type: String, index: { unique: true } },
-    type: { type: String, required: true },
-    period: { type: Number, required: true },
-    title: { type: String, required: true },
-    currentStatus: String,
-    currentStatusHistory: [String],
-    abstract: String,
-    tags: [String],
-    state: {
-      type: String,
-      required: true,
-      enum: ['preparation', 'voting', 'past'],
-    },
-    voteDate: Date,
-    submissionDate: Date,
-    lastUpdateDate: Date, // date of last dip21 history element for sorting in App
-    bioUpdateAt: Date, // Date of last dip21 changes on bundestag.io
-    subjectGroups: [String],
-    importantDocuments: [Document],
-    activities: { type: Number, default: 0 },
-    voteResults: {
-      yes: { type: Number, required: true },
-      no: { type: Number, required: true },
-      abstination: { type: Number, required: true },
-      notVoted: { type: Number, required: true },
-      decisionText: String,
-      namedVote: Boolean,
-      partyVotes: [
-        {
-          _id: false,
-          party: { type: String, required: true },
-          main: { type: String, enum: ['YES', 'NO', 'ABSTINATION', 'NOTVOTED'], required: true },
-
-          deviants: {
-            yes: { type: Number, required: true },
-            no: { type: Number, required: true },
-            abstination: { type: Number, required: true },
-            notVoted: { type: Number, required: true },
-          },
-        },
-      ],
-    },
-  },
-  { timestamps: true },
-);
-
-ProcedureSchema.methods = {
-  isVotable() {
-    return this.isVoting() || this.isCompleted();
-  },
-  isVoting() {
-    return (
-      this.currentStatus === 'Beschlussempfehlung liegt vor' ||
-      (this.currentStatus === 'Überwiesen' &&
-        this.voteDate &&
-        new Date(this.voteDate) >= new Date())
-    );
-  },
-  isCompleted() {
-    return procedureStates.COMPLETED.some(s => s === this.currentStatus || this.voteDate);
-  },
-};
-
-ProcedureSchema.index(
-  {
-    procedureId: 'text',
-    title: 'text',
-    abstract: 'text',
-    tags: 'text',
-    subjectGroups: 'text',
-  },
-  {
-    name: 'searchIndex',
-    default_language: 'german',
-    weights: {
-      title: 10,
-      abstract: 5,
-    },
-  },
-);
+import mongoose from 'mongoose';
+import ProcedureSchema from './../migrations/3-schemas/Procedure';
 
 export default mongoose.model('Procedure', ProcedureSchema);
 
+/*
 mongoose.model('Procedure').createIndexes(err => {
   if (!err) {
     Log.info('SearchIndexs for Procedures created');
@@ -95,3 +11,4 @@ mongoose.model('Procedure').createIndexes(err => {
     Log.error(JSON.stringify({ err }));
   }
 });
+*/
